fixed user experience in central activity

+++ app/src/main/java/it/polimi/mobile/design/entities/User.kt
@@ -8,5 +8,5 @@ data class User(
     val weight: Float? = null,
     val age: Int? = null,
     val gender: Gender? = null,
-    val exp: Int? = null
+    val exp: Float? = null
 )
